+++ wicket/src/lib.rs
@@ -2,19 +2,18 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! The library that is used via the technician port to initialize a rack
+//! The TUI library that is used via the technician port to initialize a rack
 //! and perform disaster recovery.
-//!
-//! This interface is a text user interface (TUI) based wizard
-//! that will guide the user through the steps the need to take
-//! in an intuitive manner.
 
 mod dispatch;
+mod events;
+mod runner;
 mod state;
 mod ui;
 mod upload;
 mod wicketd;
-mod wizard;
 
 pub use crate::dispatch::*;
-pub use crate::wizard::*;
+pub use crate::runner::*;
+pub use events::{Action, Event, InventoryEvent};
+pub use state::State;
