+++ wicket/src/state/mod.rs
@@ -4,6 +4,36 @@
 
 //! The global state manipulated by wicket.
 
-pub mod inventory;
-pub mod rack;
-pub mod status;
+mod inventory;
+mod rack;
+mod status;
+
+pub use inventory::{Component, ComponentId, Inventory, PowerState, Sp};
+pub use rack::{KnightRiderMode, RackState};
+pub use status::{ComputedLiveness, LivenessState, ServiceStatus};
+
+/// The data state of the Wizard
+///
+/// Data is not tied to any specific screen and is updated upon event receipt.
+#[derive(Debug)]
+pub struct State {
+    pub inventory: Inventory,
+    pub rack_state: RackState,
+    pub service_status: ServiceStatus,
+}
+
+impl Default for State {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl State {
+    pub fn new() -> State {
+        State {
+            inventory: Inventory::default(),
+            rack_state: RackState::new(),
+            service_status: ServiceStatus::new(),
+        }
+    }
+}
