+++ wicket/src/ui/controls/mod.rs
@@ -2,10 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::{
-    wizard::{Action, Event, State},
-    Frame,
-};
+use crate::{Action, Event, Frame, State};
 use tui::layout::Rect;
 
 /// A [`Control`] is the an item on a screen that can be selected and interacted with.
