+++ wicket/src/wicketd.rs
@@ -12,8 +12,7 @@ use tokio::time::{interval, Duration, Instant, MissedTickBehavior};
 use wicketd_client::types::RackV1Inventory;
 use wicketd_client::GetInventoryResponse;
 
-use crate::wizard::Event;
-use crate::InventoryEvent;
+use crate::{Event, InventoryEvent};
 
 const WICKETD_POLL_INTERVAL: Duration = Duration::from_millis(500);
 const WICKETD_TIMEOUT: Duration = Duration::from_millis(1000);
